+++ NEWS.md
@@ -4,6 +4,7 @@
 - copy attributes of input object x to output object y in functions filter, filtfilt, sosfilt, fftfilt
 - added ultrwin() function
 - adapted filter() to allow data and filter coefficients to be of type complex
+- adapted sosfilt() to allow data and filter coefficients to be of type complex
 
 # gsignal 0.3-2